+++ modules/Students/includes/Other_Info.inc.php
@@ -101,7 +101,6 @@ foreach ( (array) $fields_RET as $field )
 			break;
 
 		case 'exports':
-		case 'codeds':
 		case 'select':
 
 			echo _makeSelectInput( 'CUSTOM_' . $field['ID'], $field['TITLE'], 'students' );